Moved the original code under v1/, starting from scratch.
Reading some guides on raytracing (instead of trying to come up with everything
on my own, which i couldn't fully do previously (the rendered sphere ends up
skewed towards the sides of the screen, something wrong with fov calculations
probably)).

Re-starting the code (mostly) from scratch. Here is the very first step - we
draw a picture of a mix of red and green colors
(fully red on the right, fully green on the bottom).