Use Color32 for rendered frames and frame history.
This finally now results in realistic lighting, when the luminosity of the
light source is increased to high values (i.e. when using e.g. ~39*white color
(10000), instead of just white color (255)).

But there is another problem now (when using a small light with that high
luminosity) - there is pixelation in the result, where some pixels are much
lighter than other nearby pixels (in a random pattern). I tried changing the
random number generator (thinking it could be at fault here), but that didn't
help.

Also, fully removed ColorSum (it had remained in a couple places).