Fix FOV calculation for non-square image output.
Vertical/horizontal FOV proportion calc was wrong. There should only be one fov
(either vertical or horizontal), not two. And then what changes (based on
screen aspect ratio, i.e. if aspect ratio is not 1:1) are the dimensions of the
imaginary viewing plane only, not the fov.
The issue was noticeable under aspect ratio 3:2, with FOV ~= 90 (not really
noticeable with FOV ~= 40).