Glass material works well (i think).
This was achieved by fixing some issues in reflection, refraction. Also in
ray_distance_to_sphere() calculation: it previously messed up refraction for
some rays - sometimes it thought that when a ray refracts from the surface of a
sphere to inside the sphere - that it hits the sphere at its origination point,
therefore returning distance to ray/sphere boundary intersection as ~0 and the
raytracer would treat that as if the ray didn't hit the sphere at all.